Move setting of alerts, user_roles, and certified into beforeFilter()
The logic of using beforeRender was "What if the results of a form being
processed changed what these variables would end up being, such as a
login form making user_roles non-null?" But upon reconsideration, I
don't think there's any case that a form like that doesn't cause a
redirect, which would restart the loading cycle and hit beforeRender()
again. So this should be fine, plus it would allow these variables to
show up in the DebugKit toolbar, which isn't the case when they're set
in beforeRender().